pyproject.toml
[tool.poetry]
name = "character-dataset"
version = "0.1.0"
description = "repo containing characters for our dataset + scripts for preprocessing"
authors = ["Joshua Fitzmaurice <jfitz02@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
pandas = "^2.2.3"
scikit-image = "^0.24.0"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
numpy = "^2.1.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
regex = "^2024.11.6"
pyyaml = "^6.0.2"
opencv-python = "^4.10.0.84"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
